app-api-pi0-receiver/src/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-rfm9x==2.2.17
adafruit-circuitpython-typing==1.11.2
adafruit-extended-bus==1.0.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
azure-core==1.32.0
azure-storage-blob==12.24.0
binho-host-adapter==0.1.6
blinker==1.9.0
board==1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pycparser==2.22
pyftdi==0.56.0
pyserial==3.5
pyusb==1.2.1
requests==2.32.3
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
six==1.17.0
sysv-ipc==1.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
